requirement.txt
python-dotenv
requests
pandas
beautifulsoup4
SQLAlchemy
mysqlclient
PyMySQL

